changed userInfo type to NSDictionary*
git-svn-id: svn://orca.physics.unc.edu/Orca@11823 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/HV/XTR6/ORXTR6Model.h
@@ -110,7 +110,7 @@
 
 #pragma mark ***Delegate Methods (IP)
 - (void) netsocketConnected:(NetSocket*)inNetSocket;
-- (void) netsocket:(NetSocket*)inNetSocket dataAvailable:(unsigned)inAmount;
+- (void) netsocket:(NetSocket*)inNetSocket dataAvailable:(NSUInteger)inAmount;
 - (void) netsocketDisconnected:(NetSocket*)inNetSocket;
 - (BOOL) isConnected;
